change all names from skyskii to skyscii. Update assembly target and references
