+++ Skyscii.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.572
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Skyscii", "Skyscii\Skyscii.csproj", "{58C7ED11-46CA-4A75-BDE6-6442D05219CC}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{58C7ED11-46CA-4A75-BDE6-6442D05219CC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{58C7ED11-46CA-4A75-BDE6-6442D05219CC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{58C7ED11-46CA-4A75-BDE6-6442D05219CC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{58C7ED11-46CA-4A75-BDE6-6442D05219CC}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D224C7BC-92C7-45DF-B325-6CC369B72A5F}
+	EndGlobalSection
+EndGlobal
