+++ Skyscii/ITargetableObject.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Skyskii
+namespace Skyscii
 {
     interface ITargetableObject
     {
