+++ Skyscii/Item.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Skyskii
+namespace Skyscii
 {
     // consumable item logic can go here
     class Item : ITargetableObject
