+++ Skyscii/Program.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Skyskii
+namespace Skyscii
 {
     class Program
     {
