+++ Skyscii/Room.cs
@@ -2,9 +2,9 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Skyskii
+namespace Skyscii
 {
-    class Equippable : Item
+    class Room : ISearchable, ITargetableObject
     {
     }
 }
