+++ Skyscii/SentientStuff/AI.cs
