+++ Skyscii/SentientStuff/ExperienceBar.cs
