+++ Skyscii/SentientStuff/Sentient.cs
