+++ Skyscii/SentientStuff/StatBar.cs
