+++ Skyscii/SentientStuff/Stats.cs
