--- Skyskii.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27703.2000
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Skyskii", "Skyskii\Skyskii.csproj", "{2A800700-DAFD-4373-B12A-5C7461B0B749}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2A800700-DAFD-4373-B12A-5C7461B0B749}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2A800700-DAFD-4373-B12A-5C7461B0B749}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2A800700-DAFD-4373-B12A-5C7461B0B749}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2A800700-DAFD-4373-B12A-5C7461B0B749}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {F8128DC0-B7C3-4576-B6F0-FE76ADB2ECB3}
-	EndGlobalSection
-EndGlobal
